Move contents of src/lib/ to src/ and delete lib/. This is becuase lib/ is a part of micropython's path, so storing modules in lib/ was causing a silent import issue. The module imports (in the format lib.<module>) looked fine to pylint on the development machine, but micropython couldn't find them because the modules were in the sys.path, so they were importable by module name only.
