--- controller/src/lib/utility.py
@@ -1,3 +1,0 @@
-from machine import Pin # type: ignore pylint: disable=E0401
-
-onboardLed = Pin("LED", Pin.OUT)
