+++ controller/src/lightDemo.py
@@ -1,6 +1,6 @@
 # lightDemo.py
 
-from lib.lightingUnit import LightingUnit
+from lightingUnit import LightingUnit
 from time import sleep
 from random import randint
