+++ controller/src/main.py
@@ -1,5 +1,5 @@
 # main.py
 
 # Uncomment the following for an LED demo (connect WS2812's data line to GP4)
-from lib import lightDemo
+import lightDemo
 lightDemo.run()
